requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
geopandas
contextily
folium
torch
torchvision
tqdm
pyyaml
python-dotenv
sentinelhub
Pillow
joblib
statsmodels
shap
opencv-python
ipykernel
jupyter
notebook
nbconvert



